Re-implement Z-Score stats page in the Mojo frontend
The Z-Score page /zscore.*.html shows the calculated Z-Scores for all players with ascensions recorded online.
Modify original zscore() code slightly to avoid use of state variables, instead the harmonic_number()
subroutine takes two arguments and returns a tuple of two scalars as result, one of these is fed back in to the
next call to the subroutine. Calculation is slightly slow, consider storing pre-calculated scores in the DB.

Signed-off-by: Jo