Shrink container images and create podman init scripts
Ideally the system should still work with docker.
This needs to be tested. Some things are a little different.
With podman, the three containers live in a pod and share
a local network. Rootless container is used.
The initialisation of configuration is still a little messy,
this makes maintenance tricky. Definitely can be improved.

init.sh in the top level directory changes working directory,
in turn, to postgres/, cpan/, feeder/ and mojo/, calling
./init.sh, ./build.sh and ./run.sh in each (except cpan/,
where only ./build.sh is required).

Build system has been modified so that the cpan modules feeder
and mojo depend upon are built in a separate image and then
copied over to a minimal system with just a few binary libaries
as dependencies, the result is 50-60 MB images for the runtime
containers.

Signed-off-by: Joanna/aoei